Import squirrel_2.1.1_stable.tar.gz dated 2006-08-20, dos2unix'ed.

+++ sqstdlib/sqstdstream.cpp
@@ -163,7 +163,7 @@ SQInteger _stream_writen(HSQUIRRELVM v)
 	case 'f': {
 		float f;
 		sq_getfloat(v, 2, &tf);
-		f = tf;
+		f = (float)tf;
 		self->Write(&f, sizeof(float));
 			  }
 		break;
